+++ straxen/__init__.py
@@ -1,12 +1,8 @@
 # mypy: disable-error-code="no-redef"
-__version__ = "3.0.2"
+__version__ = "3.1.0"
 
 from utilix import uconfig
 from .common import *
-
-# contexts.py below
-from .corrections_services import *
-from .get_corrections import *
 
 from .itp_map import *
 from .matplotlib_utils import *