+++ straxen/contexts.py
@@ -38,9 +38,21 @@ common_opts: Dict[str, Any] = dict(
 common_config = dict(
     n_tpc_pmts=straxen.n_tpc_pmts,
     n_top_pmts=straxen.n_top_pmts,
-    gain_model="cmt://to_pe_model?version=ONLINE&run_id=plugin.run_id",
-    gain_model_nv="cmt://to_pe_model_nv?version=ONLINE&run_id=plugin.run_id",
-    gain_model_mv="cmt://to_pe_model_mv?version=ONLINE&run_id=plugin.run_id",
+    gain_model=(
+        "list-to-array://xedocs://pmt_area_to_pes"
+        "?as_list=True&sort=pmt&detector=tpc"
+        "&run_id=plugin.run_id&version=ONLINE&attr=value"
+    ),
+    gain_model_nv=(
+        "list-to-array://xedocs://pmt_area_to_pes"
+        "?as_list=True&sort=pmt&detector=neutron_veto"
+        "&run_id=plugin.run_id&version=ONLINE&attr=value"
+    ),
+    gain_model_mv=(
+        "list-to-array://xedocs://pmt_area_to_pes"
+        "?as_list=True&sort=pmt&detector=muon_veto"
+        "&run_id=plugin.run_id&version=ONLINE&attr=value"
+    ),
     channel_map=immutabledict(
         # (Minimum channel, maximum channel)
         # Channels must be listed in a ascending order!
@@ -55,40 +67,42 @@ common_config = dict(
         nveto=(2000, 2119),
         nveto_blank=(2999, 2999),
     ),
-    # Clustering/classification parameters
-    # Event level parameters
-    fdc_map=(
-        "itp_map://"
-        "resource://"
-        "cmt://"
-        "format://fdc_map_{alg}"
-        "?alg=plugin.default_reconstruction_algorithm"
-        "&version=ONLINE"
-        "&run_id=plugin.run_id"
-        "&fmt=binary"
-        "&scale_coordinates=plugin.coordinate_scales"
-    ),
-    z_bias_map=(
-        "itp_map://"
-        "resource://"
-        "XnT_z_bias_map_chargeup_20230329.json.gz?"
-        "fmt=json.gz"
-        "&method=RegularGridInterpolator"
-    ),
 )
-# these are placeholders to avoid calling cmt with non integer run_ids. Better solution pending.
-# s1, s2 and fd corrections are still problematic
 
 
-def xenonnt(cmt_version="global_ONLINE", xedocs_version=None, _from_cutax=False, **kwargs):
+@strax.Context.add_method
+def apply_xedocs_configs(context: strax.Context, db="straxen_db", **kwargs) -> None:
+    import xedocs
+
+    if isinstance(db, str):
+        func = getattr(xedocs.databases, db)
+        db_kwargs = straxen.filter_kwargs(func, kwargs)
+        db = func(**db_kwargs)
+
+    filter_kwargs = {k: v for k, v in kwargs.items() if k in db.context_configs.schema.__fields__}
+
+    docs = db.context_configs.find_docs(**filter_kwargs)
+
+    global_config = {doc.config_name: doc.value for doc in docs}
+
+    if len(global_config):
+        context.set_config(global_config)
+        context.set_context_config({"xedocs_version": filter_kwargs["version"]})
+    else:
+        warnings.warn(
+            f"Could not find any context configs matchin {filter_kwargs}",
+            RuntimeWarning,
+            stacklevel=2,
+        )
+
+
+def xenonnt(xedocs_version="global_ONLINE", _from_cutax=False, **kwargs):
     """XENONnT context."""
-    if not _from_cutax and cmt_version != "global_ONLINE":
+    if not _from_cutax and xedocs_version != "global_ONLINE":
         warnings.warn("Don't load a context directly from straxen, use cutax instead!")
-    st = straxen.contexts.xenonnt_online(**kwargs)
-    st.apply_cmt_version(cmt_version)
 
-    if xedocs_version is not None:
-        st.apply_xedocs_configs(version=xedocs_version, **kwargs)
+    st = straxen.contexts.xenonnt_online(**kwargs)
+    st.apply_xedocs_configs(version=xedocs_version, **kwargs)
 
     return st