+++ straxen/plugins/events_mv/events_mv.py
@@ -54,5 +54,5 @@ class muVETOEvents(nVETOEvents):
     def get_window_size(self):
         return self.event_left_extension_mv + self.event_resolving_time_mv + 1
 
-    def compute(self, hitlets_mv, start, end):
-        return super().compute(hitlets_mv, start, end)
+    def compute(self, hitlets_mv):
+        return super().compute(hitlets_mv)