+++ straxen/plugins/hitlets_mv/hitlets_mv.py
@@ -34,9 +34,7 @@ class muVETOHitlets(nVETOHitlets):
         track=True,
         help=(
             "Minimum hit amplitude in ADC counts above baseline. "
-            "Specify as a tuple of length n_mveto_pmts, or a number, "
-            "or a tuple like (correction=str, version=str, nT=boolean),"
-            "which means we are using cmt."
+            "Specify as a tuple of length n_mveto_pmts, or a number."
         ),
     )
 
@@ -64,7 +62,16 @@ class muVETOHitlets(nVETOHitlets):
     )
 
     gain_model_mv = straxen.URLConfig(
-        default="cmt://to_pe_model_mv?version=ONLINE&run_id=plugin.run_id",
+        default=(
+            "list-to-array://"
+            "xedocs://pmt_area_to_pes"
+            "?as_list=True"
+            "&sort=pmt"
+            "&detector=muon_veto"
+            "&run_id=plugin.run_id"
+            "&version=ONLINE"
+            "&attr=value"
+        ),
         infer_type=False,
         child_option=True,
         parent_option_name="gain_model_nv",