+++ tests/test_peaklet_processing.py
@@ -1,11 +1,13 @@
 import numpy as np
 from hypothesis import given, strategies, settings
 import hypothesis.strategies as strat
+import unittest
 import strax
 from strax.testutils import fake_hits
 import straxen
 from straxen.plugins.peaklets.peaklets import get_tight_coin
 from straxen.plugins.peaklets.peaklet_classification_som import compute_wf_attributes
+from straxen.test_utils import nt_test_run_id
 
 
 def get_filled_peaks(peak_length, data_length, n_widths):
@@ -42,21 +44,20 @@ def get_filled_peaks(peak_length, data_length, n_widths):
     return peaks
 
 
-@settings(deadline=None)
+@unittest.skipIf(not straxen.utilix_is_configured(), "No db access, cannot test!")
+@settings(deadline=None, max_examples=5)
 @given(
     strat.lists(strat.integers(min_value=0, max_value=10), min_size=8, max_size=8, unique=True),
 )
 def test_create_outside_peaks_region(time):
-    if not straxen.utilix_is_configured():
-        return
     time = strax.stable_sort(time)
     time_intervals = np.zeros(len(time) // 2, strax.time_dt_fields)
     time_intervals["time"] = time[::2]
     time_intervals["length"] = time[1::2] - time[::2]
     time_intervals["dt"] = 1
 
-    st = straxen.test_utils.nt_test_context()
-    p = st.get_single_plugin("0", "peaklets")
+    st = straxen.contexts.xenonnt_online()
+    p = st.get_single_plugin(nt_test_run_id, "peaklets")
     outside = p.create_outside_peaks_region(time_intervals, 0, np.max(time))
 
     touching = strax.touching_windows(outside, time_intervals, window=0)
@@ -66,9 +67,8 @@ def test_create_outside_peaks_region(time):
         assert np.diff(tw) == 0, "Intervals overlap although they should not!"
 
 
+@unittest.skipIf(not straxen.utilix_is_configured(), "No db access, cannot test!")
 def test_n_hits():
-    if not straxen.utilix_is_configured():
-        return
     records = np.zeros(2, dtype=strax.record_dtype())
     records["length"] = 5
     records["pulse_length"] = 5
@@ -79,7 +79,7 @@ def test_n_hits():
 
     st = straxen.contexts.xenonnt_online()
     st.set_config({"hit_min_amplitude": 1})
-    p = st.get_single_plugin("0", "peaklets")
+    p = st.get_single_plugin(nt_test_run_id, "peaklets")
     res = p.compute(records, 0, 999)
     peaklets = res["peaklets"]
     assert peaklets["n_hits"] == 3, f"Peaklet has the wrong number of hits!"